requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
exceptiongroup==1.2.0
filelock==3.13.4
gunicorn==22.0.0
h11==0.14.0
identify==2.5.35
iniconfig==2.0.0
mysqlclient==2.2.4
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
pytest==8.1.1
PyYAML==6.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
virtualenv==20.25.2
whitenoise==6.6.0
